Split up factories
