--- tests/factories/dummy/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2016 The Pycroft Authors. See the AUTHORS file.
-# This file is part of the Pycroft project and licensed under the terms of
-# the Apache License, Version 2.0. See the LICENSE file for details.
-"""
-This package contains dummy fixtures that can be used in tests. If you need
-special fixtures that are in a specific relation, please create distinct
-fixtures for these tests. If you need additional Data for a DataSet you can
-use the SuperSet class.
-"""
